+++ bb/main.clj
@@ -8,13 +8,20 @@
    [clojure.tools.logging :as log]
    [cheshire.core :as json]))
 
-;;;
 (def wc "https://wc.kohhoh.jp")
 (def cookie "cookie.txt")
 (def admin    (System/getenv "WC_LOGIN"))
 (def password (System/getenv "WC_PASSWORD"))
 (def users (atom nil))
-;;;
+
+(pods/load-pod 'org.babashka/mysql "0.1.1")
+(require '[pod.babashka.mysql :as mysql])
+(def db {:dbtype   "mysql"
+         :host     "localhost"
+         :port     3306
+         :dbname   "withings"
+         :user     (System/getenv "MYSQL_USER")
+         :password (System/getenv "MYSQL_PASSWORD")})
 
 ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
 ;;; utils
@@ -102,16 +109,6 @@
 
 ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
 ;; save meas
-
-(pods/load-pod 'org.babashka/mysql "0.1.1")
-(require '[pod.babashka.mysql :as mysql])
-(def db {:dbtype   "mysql"
-         :host     "localhost"
-         :port     3306
-         :dbname   "withings"
-         :user     (System/getenv "MYSQL_USER")
-         :password (System/getenv "MYSQL_PASSWORD")})
-
 (defn meas->float
   [{:keys [value unit]}]
   (* value (pow 10 unit)))
@@ -177,33 +174,72 @@
   (login)
   (refresh-all!)
   (get-meas-one 51 "2022-12-20")
-  ;; this must be error. `invalid user` or `user does not exist`.
+  ;; FIXME: this must be error.
+  ;;        `invalid user` or `user does not exist`.
   (get-meas-one 17 "2022-12-01")
   :rcf)
 
-;; function name init-db?
-(defn get-meas-all
-  [date]
+(defn get-save-meas-all
+  "Get all user meas since `lastupdate` via wc.kohhoh.jp,
+   save them in `withings.meas` table."
+  [lastupdate]
   (login)
   (refresh-all!)
-  ;; should valid users only
   (reset! users (fetch-users true))
   (doseq [{:keys [id]} @users]
-    (save-meas! id (get-meas-one id date))))
+    (save-meas! id (get-meas-one id lastupdate))))
 
 (comment
-  ;; should through if invalid id given
-  ;; (save-meas! 17 (get-meas-one 17 "2022-12-01"))
   (save-meas! 27 (get-meas-one 27 "2022-12-20"))
   (delete-all!)
-  (get-meas-all "2022-12-10")
-  ; clojure.lang.ExceptionInfo: babashka.curl: status 400 withings-cache /Users/hkim/clojure/withings-cache/src/withings_cache.clj:34:3
+  (get-save-meas-all "2022-12-10")
   :rcf)
 
-;; % bb -m main
-(defn -main
+(defn init-meas
+  "initialize meas table"
   [& args]
   (delete-all!)
   (if (nil? args)
-    (get-meas-all "2022-09-01")
-    (get-meas-all (first args))))
+    (get-save-meas-all "2022-09-01")
+    (get-save-meas-all (first args))))
+
+(comment
+  (init-meas)
+  :rcf)
+
+;;; updating
+;;; first delete, then add to avoid data duplications.
+(defn delete-meas-since
+  "delete meas from `date`."
+  [date]
+  (log/debug "delete-meas-since" date)
+  (mysql/execute!
+   db
+   ["delete from meas where created >= ?" date]))
+
+(defn update-meas-since
+  "deleting meas from date, then
+   fetch meas and save them."
+  [date]
+  (log/debug "update-meas-since")
+  (delete-meas-since date)
+  (get-save-meas-all date))
+
+(defn update-meas-today
+  "updating today's meas."
+  []
+  (let [today (str/trim-newline(:out (sh "date" "+%F")))]
+    (log/debug "update-meas-today today:" today)
+    (update-meas-since today)))
+
+(comment
+  (delete-meas-since "2022-12-20")
+  (update-meas-since "2022-12-20")
+  (update-meas-today)
+  :rcf)
+
+(defn -main
+  [& args]
+  (if (nil? args)
+    (update-meas-today)
+    (update-meas-since (first args))))